refactor folders to fit in design system suggestion
