+++ designSystem/atoms/Checkbox.tsx
@@ -1,6 +1,6 @@
 import React, { FC, forwardRef, HTMLAttributes } from 'react'
 import classNames from 'classnames'
-import InputLabel from '@components/atoms/InputLabelRadix'
+import InputLabel from 'designSystem/atoms/InputLabelRadix'
 
 export type CheckboxStyle = "standard" | "toggle"
 export type CheckboxProps = {
