+++ designSystem/forms/WrapperRadioGroup.tsx
@@ -1,9 +1,9 @@
 import React, { FC, JSXElementConstructor } from 'react'
 import { Controller, Control } from 'react-hook-form'
-import { RadioButton, RadioProps, RadioStyle } from '@components/atoms/RadioButton'
+import { RadioButton, RadioProps, RadioStyle } from 'designSystem/atoms/RadioButton'
 import classNames from 'classnames'
-import Tip, { TipProps } from '@components/helpers/Tip'
-import InputLabel from '@components/atoms/InputLabelRadix'
+import Tip, { TipProps } from '@components/molecules/Tip'
+import InputLabel from 'designSystem/atoms/InputLabelRadix'
 
 
 export type RadioGroupProps = {
