+++ designSystem/forms/WrapperSelect.tsx
@@ -1,8 +1,8 @@
 import React, { FC, JSXElementConstructor } from 'react'
 import { Controller, Control } from 'react-hook-form'
-import SelectRadix, {SelectRadixProps} from '@atoms/SelectRadix'
-import InputLabel from '@atoms/InputLabelRadix'
-import Tip from '@components/helpers/Tip'
+import SelectRadix, {SelectRadixProps} from 'designSystem/atoms/SelectRadix'
+import InputLabel from 'designSystem/atoms/InputLabelRadix'
+import Tip from '@components/molecules/Tip'
 
 export type WrapperSelectProps = {
   label?: string | null
