+++ pages/forms/buttons-article.tsx
@@ -1,9 +1,9 @@
-import LayoutContainerSide from '@components/layout/LayoutContainerSide'
-import Breadcrumbs from '@components/layout/Breadcrumbs'
-import Heading from '@components/layout/Heading'
-import Paragraph from '@components/layout/Paragraph'
+import LayoutContainerSide from 'designSystem/layout/LayoutContainerSide'
+import Breadcrumbs from 'designSystem/layout/Breadcrumbs'
+import Heading from 'designSystem/layout/Heading'
+import Paragraph from 'designSystem/layout/Paragraph'
 import { FC } from 'react'
-import Section from '@components/layout/Section'
+import Section from 'designSystem/layout/Section'
 
 // data for Breadcrumbs
 const crumbs = [
