+++ pages/forms/select-article.tsx
@@ -1,11 +1,11 @@
-import LayoutContainerSide from '@components/layout/LayoutContainerSide'
-import Breadcrumbs from '@components/layout/Breadcrumbs'
-import Heading from '@components/layout/Heading'
-import Paragraph from '@components/layout/Paragraph'
+import LayoutContainerSide from 'designSystem/layout/LayoutContainerSide'
+import Breadcrumbs from 'designSystem/layout/Breadcrumbs'
+import Heading from 'designSystem/layout/Heading'
+import Paragraph from 'designSystem/layout/Paragraph'
 import { FC } from 'react'
-import Section from '@components/layout/Section'
-import WrapperSelect from '@components/controlled-wrappers/WrapperSelect'
-import Divider from '@components/layout/Divider'
+import Section from 'designSystem/layout/Section'
+import WrapperSelect from '@components/forms/WrapperSelect'
+import Divider from 'designSystem/layout/Divider'
 import { useForm } from 'react-hook-form'
