+++ pages/forms/test-form.tsx
@@ -1,24 +1,24 @@
-import LayoutContainerSide from '@components/layout/LayoutContainerSide'
-import Breadcrumbs from '@components/layout/Breadcrumbs'
-import Heading from '@components/layout/Heading'
-import Paragraph from '@components/layout/Paragraph'
+import LayoutContainerSide from 'designSystem/layout/LayoutContainerSide'
+import Breadcrumbs from 'designSystem/layout/Breadcrumbs'
+import Heading from 'designSystem/layout/Heading'
+import Paragraph from 'designSystem/layout/Paragraph'
 import { FC, useEffect, useState } from 'react'
 import { useForm } from 'react-hook-form'
-import WrapperInput from '@components/controlled-wrappers/WrapperInput'
-import HelpMessage from '@components/helpers/HelpMessage'
-import WrapperRadioGroup from '@components/controlled-wrappers/WrapperRadioGroup'
-import Section from '@components/layout/Section'
-import WrapperCheckbox from '@components/controlled-wrappers/WrapperCheckbox'
-import Tip from '@components/helpers/Tip'
-import Label from '@components/atoms/InputLabelRadix'
-import TemplateGeneric from '@components/products/producta/TemplateGeneric'
-import UpdateDocState from '@components/UpdateDocState'
-import WrapperTextArea from '@components/controlled-wrappers/WrapperTextArea'
-import WrapperSelect from '@components/controlled-wrappers/WrapperSelect'
-import DatePick from '@atoms/DatePick'
-import Divider from '@components/layout/Divider'
-import BlankSpace from '@components/layout/BlankSpace'
-import Accordion from '@components/atoms/Accordion'
+import WrapperInput from '@components/forms/WrapperInput'
+import HelpMessage from '@components/molecules/HelpMessage'
+import WrapperRadioGroup from '@components/forms/WrapperRadioGroup'
+import Section from 'designSystem/layout/Section'
+import WrapperCheckbox from '@components/forms/WrapperCheckbox'
+import Tip from '@components/molecules/Tip'
+import Label from 'designSystem/atoms/InputLabelRadix'
+import TemplateGeneric from '@components/templates/producta/TemplateGeneric'
+import UpdateDocState from 'utils/UpdateDocState'
+import WrapperTextArea from '@components/forms/WrapperTextArea'
+import WrapperSelect from '@components/forms/WrapperSelect'
+import DatePick from 'designSystem/atoms/DatePick'
+import Divider from 'designSystem/layout/Divider'
+import BlankSpace from 'designSystem/layout/BlankSpace'
+import Accordion from 'designSystem/atoms/Accordion'
 
 // data for Breadcrumbs
 const crumbs = [
