+++ tsconfig.json
@@ -17,16 +17,16 @@
 
     "baseUrl": "./",
     "paths": {
-      "@components/*": ["components/*"],
-      "@atoms/*": ["components/atoms/*"],
-      "@molecules/*": ["components/molecules/*"],
+      "@components/*": ["designSystem/*"],
+      "@atoms/*": ["designSystem/atoms/*"],
+      "@molecules/*": ["designSystem/molecules/*"],
       "@layouts/*": ["layouts/*"],
       "@styles/*": ["styles/*"],
-      "@helpers/*": ["components/helpers/*"],
+      "@helpers/*": ["designSystem/molecules/*"],
       "@forms/*": ["pages/forms/*"]
 
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "components/layout/Breadcrumbs.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "designSystem/layout/Breadcrumbs.tsx"],
   "exclude": ["node_modules"]
 }
